feature: move Header from Shared folder to Widget
