+++ src/shared/loginButton/ui/loginButton.tsx
@@ -3,15 +3,24 @@ import { useTranslation } from "react-i18next"
 import { useNavigate } from "react-router-dom"
 import { ROUTES } from "../../../app/system/routes/constants"
 
-const LoginButton = (): JSX.Element | null => {
+interface LoginButtonProps {
+  closeExpandedNavbar: () => void
+}
+
+const LoginButton = ({ closeExpandedNavbar }: LoginButtonProps): JSX.Element | null => {
   const { t } = useTranslation("translation", { keyPrefix: "Navbar" })
   const navigate = useNavigate()
 
   const onClickHandler = () => {
     navigate(ROUTES.login.path)
+    closeExpandedNavbar()
   }
 
-  return <Nav.Link onClick={onClickHandler}>{t("logIn")}</Nav.Link>
+  return (
+    <Nav.Link onClick={onClickHandler} className="text-center">
+      {t("logIn")}
+    </Nav.Link>
+  )
 }
 
 export default LoginButton
